Directories refactor
