--- examples/explorer/src/components/BlockInfo/blockInfo.scss
@@ -1,13 +1,0 @@
-@import '../generalStyling.scss';
-
-.blockInfo {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.blockInfoContainer {
-  width: 80%;
-  padding: 0.5%;
-}
