--- examples/explorer/src/components/Cards/statusCard.scss
@@ -1,5 +1,0 @@
-@import '../generalStyling.scss';
-
-.statusCard {
-  width: 20%;
-}
