+++ examples/explorer/src/components/Commons/Loading/loading.scss
