--- examples/explorer/src/components/Search/search.scss
@@ -1,13 +1,0 @@
-@import '../generalStyling.scss';
-
-.searchContainer {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.searchForm {
-  width: 60%;
-  display: flex;
-  flex-direction: row;
-}
