+++ examples/explorer/src/components/Status/Cards/statusCard.scss
@@ -1,0 +1,5 @@
+@import '../../generalStyling.scss';
+
+.statusCard {
+  width: 20%;
+}
