--- examples/explorer/src/components/StatusInfo/statusInfo.scss
@@ -1,7 +1,0 @@
-@import '../generalStyling.scss';
-
-.cardsContainer {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-evenly;
-}
