--- examples/explorer/src/components/TransactionInfo/transactionInfo.scss
@@ -1,8 +1,0 @@
-@import '../generalStyling.scss';
-
-.transactionInfo {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
