--- examples/explorer/src/components/TransactionInput/transactionInput.scss
@@ -1,7 +1,0 @@
-@import '../generalStyling.scss';
-
-.transactionInput {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-}
