--- examples/explorer/src/components/TransactionInputsOutputs/transactionInputsOutputs.scss
@@ -1,16 +1,0 @@
-@import '../generalStyling.scss';
-
-.transactionInputsOutputs {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  width: 100%;
-}
-
-.column {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  width: 90%;
-  padding: 10px;
-}
