--- examples/explorer/src/components/TransactionOutput/transactionOutput.scss
@@ -1,8 +1,0 @@
-@import '../generalStyling.scss';
-
-.transactionOutput {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  width: 100%;
-}
