--- examples/explorer/src/components/TransactionSearchResult/TransactionSearchResult.jsx
@@ -1,35 +1,0 @@
-import React from 'react';
-
-import graphql from 'babel-plugin-relay/macro';
-import { QueryRenderer } from 'react-relay';
-import environment from '../../graphql/environment';
-import TransactionInfo from '../TransactionInfo/TransactionInfo';
-import Loading from '../commons/Loading/Loading';
-
-const TransactionSearchResult = ({ id }) => (
-  <QueryRenderer
-    environment={environment}
-    query={graphql`
-      query TransactionSearchResultQuery($id: String!) {
-        transaction(id: $id) {
-          id
-          ...TransactionInfo_transaction
-        }
-      }
-    `}
-    variables={{ id }}
-    render={response => {
-      const { error, props } = response;
-      if (error) {
-        return <div>Error!</div>;
-      }
-      if (!props) {
-        return <Loading />;
-      }
-
-      return <TransactionInfo {...props} />;
-    }}
-  />
-);
-
-export default TransactionSearchResult;
