+++ examples/explorer/src/components/Transactions/TransactionInfo/transactionInfo.scss
@@ -1,0 +1,8 @@
+@import '../../generalStyling.scss';
+
+.transactionInfo {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
