+++ examples/explorer/src/components/Transactions/TransactionInput/transactionInput.scss
@@ -1,0 +1,7 @@
+@import '../../generalStyling.scss';
+
+.transactionInput {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+}
