+++ examples/explorer/src/components/Transactions/TransactionInputsOutputs/TransactionInputsOutputs.jsx
