+++ examples/explorer/src/components/Transactions/TransactionOutput/transactionOutput.scss
@@ -1,0 +1,8 @@
+@import '../../generalStyling.scss';
+
+.transactionOutput {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  width: 100%;
+}
